fixed FriendsList and UserService
