+++ client/src/components/FriendsList/FriendsList.jsx
@@ -1,104 +1,43 @@
-import React, { useState } from 'react'
+import React, { Fragment } from 'react'
 import {
-  Button, Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
   Grid,
-  Slide,
   Typography
 } from '@material-ui/core'
 import PropTypes from 'prop-types'
-import { connect } from 'react-redux'
 import useStyles from './friendsListStyles'
-import HighlightOffIcon from '@material-ui/icons/HighlightOff'
-import MailOutlineIcon from '@material-ui/icons/MailOutline'
-import Tooltip from '@material-ui/core/Tooltip'
-import IconButton from '@material-ui/core/IconButton'
-import { deleteFriend } from '../../actions/auth'
-import { getAvatarLink } from '../../utils/helpers/imageLinkHelpers'
-import Tile from '../Tile/Tile'
+import FriendsListItem from './FriendsListItem/FriendsListItem'
 
-const Transition = React.forwardRef(function Transition (props, ref) {
-  return <Slide direction='up' ref={ref} {...props} />
-})
-
-const FriendsList = ({ user, deleteFriend }) => {
+const FriendsList = ({ friends, requests }) => {
   const classes = useStyles()
-  const [openDialog, setOpenDialog] = useState(false)
-  const { friends } = user
 
-  const handleModal = () => {
-    setOpenDialog(!openDialog)
+  const fieldComponents = components => {
+    if (friends) {
+      return components.map(friend => <FriendsListItem friend={friend} key={friend.avatar.id}/>)
+    } else {
+      return components.map(request => <FriendsListItem request={request} key={request.id}/>)
+    }
   }
 
-  const handleModalDelete = friendUsername => {
-    handleModal()
-    deleteFriend(friendUsername)
-  }
-
-  const friendsList = friends.map(friend =>
-    <Grid item sm={5} className={classes.gridItem} key={friend.avatar.id}>
-      <Tile imageSrc={getAvatarLink(friend.avatar)} />
-      <div className={classes.friendInfo}>
-        <p className={classes.userName}>{friend.firstName} {friend.lastName}</p>
-        <div>
-          <Tooltip title='Send message'>
-            <IconButton color='primary' aria-label='Send message'>
-              <MailOutlineIcon />
-            </IconButton>
-          </Tooltip>
-          <Tooltip title='Remove friend'>
-            <IconButton color='secondary' onClick={handleModal} aria-label='Remove friend'>
-              <HighlightOffIcon />
-            </IconButton>
-          </Tooltip>
-          <Dialog
-            open={openDialog}
-            TransitionComponent={Transition}
-            keepMounted
-            onClose={handleModal}
-          >
-            <DialogTitle id='alert'>Delete friend</DialogTitle>
-            <DialogContent>
-              <DialogContentText>
-                Are you sure you want to delete friend?
-              </DialogContentText>
-            </DialogContent>
-            <DialogActions>
-              <Button variant='contained' color='primary' onClick={handleModal}>
-                Cancel
-              </Button>
-              <Button variant='contained' color='secondary' onClick={() => handleModalDelete(friend.username)}>
-                Delete
-              </Button>
-            </DialogActions>
-          </Dialog>
-        </div>
-      </div>
-    </Grid>
-  )
+  const content = (friends) ? fieldComponents(friends) : fieldComponents(requests)
 
   return (
     <div className={classes.container}>
       <Typography className={classes.header} variant='subtitle1' component='div'>
-        Friends <span className={classes.count}>{friends.length}</span>
+        { friends
+          ? <Fragment>Friends <span className={classes.count}>{friends.length}</span></Fragment>
+          : <Fragment>Friend requests <span className={classes.count}>{requests.length}</span></Fragment>
+        }
       </Typography>
       <Grid className={classes.gridContainer} container>
-        {friendsList}
+        {content}
       </Grid>
     </div>
   )
 }
 
 FriendsList.propTypes = {
-  user: PropTypes.object.isRequired,
-  deleteFriend: PropTypes.func.isRequired
+  friends: PropTypes.array,
+  requests: PropTypes.array,
 }
 
-const mapStateToProps = state => ({
-  user: state.auth.user
-})
-
-export default connect(mapStateToProps, { deleteFriend })(FriendsList)
+export default FriendsList
