+++ client/src/components/FriendsList/FriendsListItem/friendsListItemStyles.js
@@ -1,24 +1,6 @@
 import { makeStyles } from '@material-ui/core/styles'
 
 const useStyles = makeStyles(() => ({
-  container: {
-    textAlign: 'left'
-  },
-  header: {
-    color: '#1c1e21',
-    padding: '8px 10px',
-    fontWeight: 900,
-    borderBottom: '1px solid #dddfe2',
-    paddingRight: 24,
-    backgroundColor: '#f5f6f7'
-  },
-  count: {
-    color: '#90949c',
-    fontWeight: 500
-  },
-  gridContainer: {
-    padding: '8px 10px'
-  },
   gridItem: {
     display: 'flex',
     border: '1px solid #e9ebee',
@@ -32,7 +14,7 @@ const useStyles = makeStyles(() => ({
     width: '100%',
     padding: '0 10px'
   },
-  friendName: {
+  userName: {
     color: '#000',
     fontWeight: 600,
     margin: 0,
