+++ client/src/components/FriendsList/friendsListStyles.js
@@ -19,25 +19,6 @@ const useStyles = makeStyles(() => ({
   gridContainer: {
     padding: '8px 10px'
   },
-  gridItem: {
-    display: 'flex',
-    border: '1px solid #e9ebee',
-    margin: '8px auto',
-    minWidth: 370
-  },
-  friendInfo: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    width: '100%',
-    padding: '0 10px'
-  },
-  userName: {
-    color: '#000',
-    fontWeight: 600,
-    margin: 0,
-    wordBreak: 'break-word'
-  }
 
 }))
